# openapi_generated.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /languages:
    get:
      summary: Get Languages
      operationId: get_languages
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/LanguageModel'
                type: array
                title: Response Get Languages
  /translations:
    get:
      summary: Get Translations
      operationId: get_translations
      parameters:
      - name: language
        in: query
        required: false
        schema:
          title: Language
          type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TranslationModel'
                title: Response Get Translations
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /hello:
    get:
      summary: Root
      operationId: root_hello_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LanguageModel:
      properties:
        alias:
          type: string
          title: Alias
        name_en:
          type: string
          title: Name En
        name_national:
          type: string
          title: Name National
      type: object
      required:
      - alias
      - name_en
      - name_national
      title: LanguageModel
    TranslationModel:
      properties:
        code:
          type: integer
          title: Code
        alias:
          type: string
          title: Alias
        name:
          type: string
          title: Name
        description:
          title: Description
          type: string
        language:
          type: string
          title: Language
        voices:
          items:
            $ref: '#/components/schemas/VoiceModel'
          type: array
          title: Voices
      type: object
      required:
      - code
      - alias
      - name
      - language
      - voices
      title: TranslationModel
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VoiceModel:
      properties:
        code:
          type: integer
          title: Code
        alias:
          type: string
          title: Alias
        name:
          type: string
          title: Name
        description:
          title: Description
          type: string
        is_music:
          type: boolean
          title: Is Music
      type: object
      required:
      - code
      - alias
      - name
      - is_music
      title: VoiceModel